requirements.txt
flask
flask-login
requests
RPi.GPIO
redis
peewee
flask-peewee
python-telegram-bot
PySocks